user-data-providers: conditionally compile vmware guestinfo for x86 only
