+++ sources/early-boot-config/user-data-providers/vmware-cd-rom/README.md
@@ -4,7 +4,7 @@ Current version: 0.1.0
 
 ## Introduction
 
-User data provider binary that fetches user data provided via CD-ROM to a VMWare VM.
+User data provider binary that fetches user data provided via CD-ROM to a VMware VM.
 
 ## Colophon
