+++ sources/early-boot-config/user-data-providers/vmware-cd-rom/src/main.rs
@@ -1,7 +1,7 @@
 /*!
 # Introduction
 
-User data provider binary that fetches user data provided via CD-ROM to a VMWare VM.
+User data provider binary that fetches user data provided via CD-ROM to a VMware VM.
 */
 
 use early_boot_config_provider::provider::{
