+++ sources/early-boot-config/user-data-providers/vmware-guestinfo/Cargo.toml
@@ -15,6 +15,9 @@ serde = { version = "1", features = ["derive"] }
 serde_plain = "1"
 snafu = "0.8"
 early-boot-config-provider = { path = "../../early-boot-config-provider", version = "0.1" }
+
+[target.'cfg(target_arch = "x86_64")'.dependencies]
+# vmw_backdoor includes x86_64 assembly, prevent it from building for ARM
 vmw_backdoor = "0.2"
 
 [build-dependencies]
